+++ src/utilities/get-config-extensions.ts
@@ -1,0 +1,8 @@
+import * as vscode from "vscode";
+
+export type GetConfigExtensionsProps = {};
+
+export function getConfigExtensions({}: GetConfigExtensionsProps) {
+  const config = vscode.workspace.getConfiguration("copyFolderContent");
+  return config.get<string[]>("include", ["**/*.{js,ts,jsx,tsx}"]);
+}