+++ src/utilities/handle-error.ts
@@ -1,0 +1,9 @@
+import * as vscode from "vscode";
+
+export type HandleErrorProps = {
+  error?: any;
+};
+
+export function handleError({ error }: HandleErrorProps) {
+  vscode.window.showErrorMessage(`Error: ${error.message}`);
+}