+++ src/utilities/handle-info.ts
@@ -1,0 +1,5 @@
+import * as vscode from "vscode";
+
+export function handleInfo(info: string) {
+  vscode.window.showInformationMessage(info);
+}